Show sub-cent values on Referrals chart Y axis and tooltips FEDEV-4264
